Extract template records functions in a separated file
